+++ src/widgets/BhandaraCheckin/store/slices/index.ts
@@ -1,3 +1,3 @@
-export * from "../../BhandaraCheckin/bhandara-checkin";
+export * from "../../BhandaraCheckin/bhandaraCheckinSlice";
 export * from "../../SectionMain/mainSectionSlice";
 export * from "../../SectionUpdateDetails/updateDetailsSectionSlice";
